% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{add}
\alias{add}
\title{Add together two numbers.}
\usage{
add(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
The sum of \code{param1} and \code{param2}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
