% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{div}
\alias{div}
\title{Divide two numbers.}
\usage{
div(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
The division of \code{param1} and \code{param2}.
}
\description{
Divide two numbers.
}
\examples{
div(1, 1)
div(10, 1)
}
