% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{isEqual}
\alias{isEqual}
\title{Compare parameter1 to parameter2 for equal.}
\usage{
isEqual(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
TRUE/FALSE.
}
\description{
Compare parameter1 to parameter2 for equal.
}
\examples{
isEqual(2, 2)
isEqual(10, 1)
}
