% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{isGreater}
\alias{isGreater}
\title{Compare parameter1 to parameter2 for greater.}
\usage{
isGreater(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
TRUE/FALSE.
}
\description{
Compare parameter1 to parameter2 for greater.
}
\examples{
isGreater(2, 1)
isGreater(10, 1)
}
