% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{isGreaterOrEqual}
\alias{isGreaterOrEqual}
\title{Compare parameter1 to parameter2 for greater or equal.}
\usage{
isGreaterOrEqual(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
TRUE/FALSE.
}
\description{
Compare parameter1 to parameter2 for greater or equal.
}
\examples{
isGreaterOrEqual(2, 1)
isGreaterOrEqual(10, 1)
}
