% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{islesser}
\alias{islesser}
\title{Compare parameter1 to parameter2 for lesser.}
\usage{
islesser(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
TRUE/FALSE.
}
\description{
Compare parameter1 to parameter2 for lesser.
}
\examples{
islesser(2, 1)
islesser(10, 1)
}
