% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{islesserOrEqual}
\alias{islesserOrEqual}
\title{Compare parameter1 to parameter2 for lesser.}
\usage{
islesserOrEqual(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
TRUE/FALSE.
}
\description{
Compare parameter1 to parameter2 for lesser.
}
\examples{
islesserOrEqual(2, 1)
islesserOrEqual(10, 1)
}
