% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{mul}
\alias{mul}
\title{Multiply two numbers.}
\usage{
mul(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
The product of \code{param1} and \code{param2}.
}
\description{
Multiply two numbers.
}
\examples{
mul(1, 1)
mul(10, 1)
}
