% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sub}
\alias{sub}
\title{Substract two numbers.}
\usage{
sub(param1, param2)
}
\arguments{
\item{param1}{A number.}

\item{param2}{A number.}
}
\value{
The difference of \code{param1} and \code{param2}.
}
\description{
Substract two numbers.
}
\examples{
sub(1, 1)
sub(10, 1)
}
